requirements.txt
Django==4.2.7
pillow==10.2.0
pydrive2==1.19
python-dotenv==1.0.1
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
djangorestframework==3.14.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
django-allauth==0.54.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.1.0
django-filter~=24.2
PyJWT~=2.8.0
typing~=3.7.4.3
attrs~=23.2.0
oauthlib~=3.2.2
flask~=3.0.2
environ~=1.0
services~=0.1.1
config~=0.5.1
django-oauth2-provider==0.2.6.1
social-auth-app-django==5.4.0
python-social-auth==0.2.21
PyDrive==1.3.1